+++ esperanto-tools-libs/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [[_TOC_]]
 ## [Unrealeased]
 ### Added
+- [SW-15494] Added an individual threadpool for each device
 - [SW-15364] Automatic doxygen documentation generation.
 - [SW-15254] CMA allocation policy updated.
 - [SW-15419] Avoid using CMA for loading kernels when not needed.